rename: change AddressSeq to AddressQueueIndex and update module references
