+++ sdk-libs/photon-api/src/models/mod.rs
@@ -18,8 +18,8 @@ pub mod account_with_optional_token_data;
 pub use self::account_with_optional_token_data::AccountWithOptionalTokenData;
 pub mod account_with_optional_token_data_v2;
 pub use self::account_with_optional_token_data_v2::AccountWithOptionalTokenDataV2;
-pub mod address_seq;
-pub use self::address_seq::AddressQueueIndex;
+pub mod address_queue_index;
+pub use self::address_queue_index::AddressQueueIndex;
 pub mod address_with_tree;
 pub use self::address_with_tree::AddressWithTree;
 pub mod closed_account_v2;
